Override tmux socket directory to be /tmp
By default it's /run/user/$UID, which is a tmpfs that is
deleted when all of the user's sessions close. This means that
after disconnecting and connecting (e.g. over ssh), old tmux
sessions are still alive but cannot be attached.

This commit forces it to be the slightly more permanent /tmp.

Note that a `programs.tmux.secureSocket` option has recently been
added to NixOS but hasn't made it into the current release channel
yet. When it does, setting it to false accomplishes the same without
the need for a `mkForce`.