Move kubernetes datadir to /var/lib/kubelet
Previously it was /var/lib/kubernetes, with a symlink
/var/lib/kubelet -> /var/lib/kubernetes in place to make the
rook-ceph operator happy (it needs to have the same kubeletdir
across all nodes).

However, it turns out kubelet sometimes resolves the symlink in
the path, resulting in it sending paths under /var/lib/kubernetes
to the csi-rbdplugin, but this path isn't mounted in the csi-rbdplugin
container.